+++ R/lambda.R
@@ -13,6 +13,8 @@
 #' @param env environment to work in.
 #' @return user defined function.
 #'
+#' @seealso \code{\link{lambda}}, \code{\link{defineLambda}}, \code{\link{named_map_builder}}
+#'
 #' @examples
 #'
 #' f <- makeFunction_se(as.name('x'), substitute({x*x}))
@@ -33,7 +35,9 @@ makeFunction_se <- function(params, body, env = parent.frame()) {
   vars <- as.character(params)
   formals <- replicate(length(vars), quote(expr = ))
   names(formals) <- vars
-  eval(call('function', as.pairlist(formals), body), env)
+  eval(call('function', as.pairlist(formals), body),
+       envir = env,
+       enclos = env)
 }
 
 
@@ -45,6 +49,9 @@ makeFunction_se <- function(params, body, env = parent.frame()) {
 #' @param ... formal parameters of function, unbound names, followed by function body (code/language).
 #' @param env environment to work in
 #' @return user defined function.
+#'
+#'
+#' @seealso \code{\link{defineLambda}}, \code{\link{makeFunction_se}}, \code{\link{named_map_builder}}
 #'
 #' @examples
 #'
@@ -59,8 +66,12 @@ makeFunction_se <- function(params, body, env = parent.frame()) {
 #' f <- lambda(x, y, x+y)
 #' f(2,4)
 #'
-#' # formula interface syntax: [~arg|arg(~arg)+] := body
-#' f <- x~y := x + 3 * y
+#' # brace interface syntax
+#' f <- x := { x^2 }
+#' f(5)
+#'
+#' # formula interface syntax: [~arg|arg(~arg)+] := { body }
+#' f <- x~y := { x + 3 * y }
 #' f(5, 47)
 #'
 #' @export
@@ -82,6 +93,8 @@ lambda <- function(..., env = parent.frame()) {
 #'
 #' @param envir environment to work in.
 #' @param name character, name to assign to (defaults to Greek lambda).
+#'
+#' @seealso \code{\link{lambda}}, \code{\link{makeFunction_se}}, \code{\link{named_map_builder}}
 #'
 #' @examples
 #'